pyproject.toml
[tool.poetry]
name = "mhast"
version = "0.1.0"
description = ""
authors = ["Ali Moeeny <alimoeeny@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
tensorflow = "^2.5.0"
Pillow = "^8.2.0"
wandb = "^0.10.33"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
